+++ src/main/java/com/victor/project/gymapp/repositories/UserRepository.java
@@ -20,10 +20,10 @@ public interface UserRepository extends CrudRepository<User, String> {
 
     // Devuelve el uuid del propietario de la temporada
     @Query("SELECT s.user.uuid FROM Season s WHERE s.id = :seasonId")
-    Optional<String> findUserUuidBySeason(@Param("seasonId") Long seasonId);
+    Optional<String> findUserUuidBySeason(@Param("seasonId") Integer seasonId);
 
     // Devuelve el uuid del propietario del registro de usuario
     @Query("SELECT r.user.uuid FROM UserRecord r WHERE r.id = :userRecordId")
-    Optional<String> findUserUuidByUserRecord(@Param("userRecordId") Long userRecordId);
+    Optional<String> findUserUuidByUserRecord(@Param("userRecordId") Integer userRecordId);
 
 }
